Tweak Marshlight appearance
Make will-o-the-wisps' foxfires match their colouring, and tweak
messaging. This should hopefully make players more scared of them.
(At least in tiles, sorry!)

Some argument for giving the marshlights a different name and/or a
fully separate monster type.

This may break save compat.

+++ crawl-ref/source/mapdef.cc
@@ -4068,7 +4068,7 @@ mons_list::mons_spec_slot mons_list::parse_mons_spec(string spec)
             }
             // Store name along with the tile.
             mspec.props[MONSTER_TILE_NAME_KEY].get_string() = tile;
-            mspec.props[MONSTER_TILE_KEY] = short(index);
+            mspec.props[MONSTER_TILE_KEY] = int(index);
         }
 
         string dbname = strip_tag_prefix(mon_str, "dbname:");
